requirements.txt
ldap3
openai
python-dotenv
langchain
streamlit
streamlit-option-menu
streamlit-cookies-manager-hotpatched
streamlit-dimensions
streamlit-pdf-viewer
streamlit-pills
PyYAML
PyMuPDF
pypdf
pdf2image
chromadb
